reduxを導入してcombineRuducersを使用するように
