+++ package.json
@@ -6,7 +6,8 @@
     "bootstrap": "4.3.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-scripts": "3.4.1"
+    "react-scripts": "3.4.1",
+    "redux": "4.0.5"
   },
   "scripts": {
     "start": "react-scripts start",
